pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-backtesting"
version = "0.1.0"
description = "ETF comparison and backtesting dashboard scaffold"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "plotly>=5.24",
  "PyYAML>=6.0",
  "streamlit>=1.44",
  "yfinance>=0.2.54",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

